Computation of QM statistics
Calculation of QM statistics depends on following inputs:
   1. RRD files for each product created by trollduction postprocessor (contains infos about creation and processing time for each timeslot)
   2. EUMETCAST daily log files (information about transfer of sat data to DWD)
   3. DWD AFD alda log files (information about transfer to DWD pytroll server via AFD)

RRD files currently track data for last 3 months. EUMETCAST daily log and AFD alda log files have to be cached because the corresponding files will be deleted after some days.

Importing EUMETCAST Daily Log

To import EUMETCAST daily log use the script "bin/import_daily_logs.py.
The corresponding configuration file is located in "config/daily_log_cfg.yml":

  # location of file database containing cached log information
  daily_log_db_filename: "/opt/pytroll/pytroll_inst/config/daily_log.db"

  # remove all entries from db that are older than x days compared to newest entry
  max_age_days: 120

  # map to assign products to relevant log entries
  prod2dailylog_records_pattern_map:
    "METEOSAT_EUROPA_GESAMT_IR108_.*":
      - {"source": "MSG_0DEG", "service": "H_SEVIRI", "satellite": "MSGx", "channel": "%", "segment": "EPI"}
      - {"source": "MSG_0DEG", "service": "H_SEVIRI", "satellite": "MSGx", "channel": "IR_108", "segment": "%"}
    "WarnappbildRGBA_.*":
      - {"source": "MSG_0DEG", "service": "H_SEVIRI", "satellite": "MSGx", "channel": "%", "segment": "EPI"}

Importing AFD Alda Log

To import AFD alda log files use the script "bin/import_afd_alda_logs.py.
The corresponding configuration file is located in "config/alda_log_cfg.yml":

  # location of file database containing cached log information
  alda_log_db_filename: "/opt/pytroll/pytroll_inst/config/afd_alda_log.db"

  # remove all entries from db that are older than x days compared to newest entry
  max_age_days: 120

  # patterns to extract timeslot from filenames listed in each line of a log file
  # first matching pattern will be used
  reader_filename_patterns: ["{dummy}EPI______-{time:%Y%m%d%H%M}-__", "{dummy}_CT___{time:%Y%m%d%H%M}_EUROPE_B____"]

  # map to assign products to relevant log entries
  # "min_count" defines how many entries should match the filter to be handled as "ok"
  prod2aldalog_records_pattern_map:
    "METEOSAT_EUROPA_GESAMT_IR108_.*":
      - {"min_count": 1, "dest_host": "oflxs551", "filename": "H-000-MSG%-EPI%"}
    "WarnappbildRGBA_.*":
      - {"min_count": 1, "dest_host": "oflxs551", "filename": "H-000-MSG%-EPI%"}
      - {"min_count": 1, "dest_host": "oflxs551", "filename": "SAFNWC_%_CT%-EPI%"}

Configuration of computation of QM stats

The configuration file is located in "config/calc_qm_stats_cfg":

  # path to configuration of EUMETCAST Daily Log
  daily_log_config: "/opt/pytroll/pytroll_inst/config/daily_log_cfg.yml"

  # path to configuration of AFD Alda Log
  alda_log_config: "/opt/pytroll/pytroll_inst/config/alda_log_cfg.yml"

  # path to directory containing RRD files created by postprocessor
  rrd_directory: "/opt/pytroll/pytroll_inst/log/rrd"

  # lists of products to calculate QM metrics for
  # steps: timedelta between two timeslots in seconds
  # output_csv_file: filename of the csv file to append result
  # allowed_process_time: max process time before handle timeslot of the product as delayed
  products:
      - {name: "METEOSAT_EUROPA_GESAMT_IR108_nqeuro3km_xx.tif", steps: 900, output_csv_file: "/opt/pytroll/pytroll_inst/log/qm_stats_ir108.csv", allowed_process_time: 60}
      - {name: "WarnappbildRGBA_pcwarnapp1500m_xx.tif", steps: 900, output_csv_file: "/opt/pytroll/pytroll_inst/log/qm_stats_warnapp.csv", allowed_process_time: 60}

Run computation of QM stats

  # calc stats for 11/2016
  /opt/pytroll/pytroll_inst/tools/calc_qm_stats.sh 2016 11

  # calc stats for previous month
  /opt/pytroll/pytroll_inst/tools/calc_qm_stats.sh

+++ dwd_extensions/qm/emc_dailylogs/repository.py
@@ -119,23 +119,32 @@ class Repository(object):
         ''' find all records in repository '''
         return self.session.query(DailyLogEntry).all()
 
-    def delete_older_than(self, days):
+    def get_oldest_allowed_date(self, days):
         ''' delete all records with reference_time older
         than *days* before newest record '''
         max_date = self.session.query(
             func.max(DailyLogEntry.reference_time)).select_from(
                 DailyLogEntry).scalar()
 
-        if max_date > datetime.utcnow():
-            print "max date in database is in future, using current date"
-            max_date = datetime.utcnow()
+        if max_date is not None:
+            if max_date > datetime.utcnow():
+                print "max date in database is in future, using current date"
+                max_date = datetime.utcnow()
 
-        max_date = max_date - timedelta(days=days)
+            max_date = max_date - timedelta(days=days)
 
-        print "deleting all records before {}".format(max_date)
+        return max_date
 
-        self.session.query(DailyLogEntry).filter(
-            DailyLogEntry.reference_time < max_date).delete()
+    def delete_older_than(self, days):
+        ''' delete all records with reference_time older
+        than *days* before newest record '''
+
+        max_date = self.get_oldest_allowed_date(days)
+        if max_date is not None:
+            print "deleting all records before {}".format(max_date)
+
+            self.session.query(DailyLogEntry).filter(
+                DailyLogEntry.reference_time < max_date).delete()
 
     def record_count(self):
         ''' return count of accouncements in repository '''
